# Conflicts:
#	ehcore/src/main/java/com/everhomes/pmtask/PmTaskController.java
#	ehcore/src/main/java/com/everhomes/pmtask/PmTaskProviderImpl.java
#	ehcore/src/main/java/com/everhomes/pmtask/PmTaskServiceImpl.java
#	ehnextgen-api/src/main/java/com/everhomes/pmtask/PmTaskProvider.java
#	ehnextgen-api/src/main/java/com/everhomes/pmtask/PmTaskService.java

+++ ehnextgen-api/src/main/java/com/everhomes/order/PayProvider.java
@@ -18,7 +18,7 @@ public interface PayProvider {
 
     PaymentUser findPaymentUserByOwner(String ownerType, Long ownerId);
 
-    PaymentAccount findPaymentAccountBySystemId(Integer systemId);
+    PaymentAccount findPaymentAccountBySystemId(Long systemId);
 
     void createPaymentOrderRecord(PaymentOrderRecord orderRecord);
